Split docs/serve generation
